FIX lr_fact is a float in learner2::get_early_stopper

+++ src/cuvnet/tools/learner2.cpp
@@ -246,7 +246,7 @@ namespace cuvnet
         int boxfilter = cfg.get("boxfilter", 1);
         int patience = cfg.get("patience", 100);
         int max_steps = cfg.get("max_steps", 0);
-        int lr_fact = cfg.get("lr_fact", 0.5f);
+        float lr_fact = cfg.get("lr_fact", 0.5f);
         LOG4CXX_WARN(g_log_learner2, "Setting up Early Stopper (watch:"<<watch
                 <<", thresh:"<< thresh
                 << ", every: "<< every
